[mvsUtils] Move ROI header from depthMap to mvsUtils
