+++ src/aliceVision/mvsUtils/ROI.hpp
